Implemented external stimulus input
1. Implemented external stimulus input (regular pulse)
2. Use -s option to specify stimulus file
3. Fixed Network::reset() - currently this method is not used, but
fixed some problems.
4.  m_summationMap in Network was used wrong. Now it is the reference
of the summationMap pointer occupied by neurons.
5. simInfo-> pSummationMap is also the reference of the summationMap
pointer occupied by neurons, and it is set properly (pointer to the
device memory in cuda version).
6. Added stimulus input configuration files.
7. Fixed parameter description files for 10x10 (Added maxFiringRate and
maxSynapsesPerNeuron)

+++ common/XmlRecorder.cpp
@@ -96,11 +96,11 @@ void XmlRecorder::compileHistories(const AllNeurons &neurons, BGFLOAT minRadius)
         for (int i = 0; i < spike_count; i++)
         {
             // compile network wide burstiness index data in 1s bins
-            int idx1 = static_cast<int>( pSpikes[i] * m_sim_info->deltaT );
+            int idx1 = static_cast<int>( static_cast<double>( pSpikes[i] ) * m_sim_info->deltaT );
             burstinessHist[idx1] = burstinessHist[idx1] + 1.0;
 
             // compile network wide spike count in 10ms bins
-            int idx2 = static_cast<int>( pSpikes[i] * m_sim_info->deltaT * 100 );
+            int idx2 = static_cast<int>( static_cast<double>( pSpikes[i] ) * m_sim_info->deltaT * 100 );
             spikesHistory[idx2] = spikesHistory[idx2] + 1.0;
         }
